+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 19.10
 
 + Change the term "nodes" to "vertices".
++ Replace 'ground' string as "VLink.FRAME" class attribute.
 + Add placeholder "HOLDER" class variable for VPoint and VLink.
 + Support MSVC compiler.
 + Support PEP 561.